feat: do not remove magic-links on use
These magic-links are one use only which is fine unless something like
an email client or other helpful tool scans the link, thereby using it
and preventing the actual user from being able to use it. In order to
avoid situations where this happens the magic-links will only expire on
the timeout, rather than use.